requirements.txt
fastapi
uvicorn
PyYAML

httpx
requests
bs4
pymysql
dbutils
apscheduler

#utils 依赖
python-consul
